backend/requirements.txt
google-cloud-storage>=2.16.0
google-cloud-videointelligence>=2.11.2
opencv-python>=4.9.0.80
fastapi>=0.112.0
uvicorn>=0.30.0
numpy>=1.26.0
pydantic>=2.7.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
firebase-admin>=6.5.0
python-multipart>=0.0.9
google-cloud-pubsub>=2.20.0
